+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0.2] - 2025-10-27
+## [0.4.0.2] - 2025-10-29
 
 ### Versions
 
-- CSC FastAPI Backend:  0.4.0.0
+- CSC FastAPI Backend:  0.4.0.1
 - CSC React Frontend:   0.4.0.2
 - CSC Grasshopper Interface: 0.4.0.0
 
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CSC Grasshopper Interface: 0.4.0.0
 
 ### Changed
+
+#### CSC FastAPI Backend
+
+- **Descriptor Computation**: Fixed Timestamps, added timestamp utility to use in all routes
 
 #### CSC React Frontend